# docs/swagger.yaml
basePath: /v1/core-services
definitions:
...
  github_com_modami_core-service_internal_dto.CancelOrderRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
  github_com_modami_core-service_internal_dto.CreateOrderRequest:
    properties:
      product_id:
        type: string
      shipping:
        $ref: '#/definitions/github_com_modami_core-service_internal_dto.CreateShippingInput'
    required:
    - product_id
    - shipping
    type: object
...
  github_com_modami_core-service_internal_dto.CreateShippingInput:
    properties:
      address:
        type: string
      district:
        type: string
      name:
        type: string
      phone:
        type: string
      province:
        type: string
      ward:
        type: string
    required:
    - address
    - district
    - name
    - phone
    - province
    - ward
    type: object
...
  github_com_modami_core-service_internal_dto.ShipOrderRequest:
    properties:
      shipping_provider:
        type: string
      tracking_code:
        type: string
    required:
    - shipping_provider
    - tracking_code
    type: object
...
  internal_adapter_handler.StandardErrorEnvelope:
    properties:
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_adapter_handler.StandardSuccessEnvelope:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Core service: catalog, orders, master data, and seller APIs for the
    Modami marketplace.'
  title: Modami Core Service API
  version: "1.0"
paths:
...
  /admin/orders:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: list all orders'
      tags:
      - Admin
  /admin/orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: get order'
      tags:
      - Admin
  /admin/orders/{id}/force-cancel:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: force cancel order'
      tags:
      - Admin
  /admin/orders/{id}/force-complete:
    put:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: 'Admin: force complete order'
      tags:
      - Admin
...
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create order (purchase)
      tags:
      - Orders
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - Orders
  /orders/{id}/confirm:
    put:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Seller confirms order
      tags:
      - Orders
  /orders/{id}/events:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Order timeline events
      tags:
      - Orders
  /orders/{id}/receive:
    put:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Buyer confirms receipt
      tags:
      - Orders
  /orders/{id}/ship:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Tracking
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_core-service_internal_dto.ShipOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Seller ships order
      tags:
      - Orders
  /orders/my-purchases:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List orders I purchased
      tags:
      - Orders
  /orders/my-sales:
    get:
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardSuccessEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_handler.StandardErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List orders I sold
      tags:
      - Orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"